fix(payroll): mark as paid gagal saat ada cicilan pinjaman jatuh tempo
Menandai payroll "paid" mengembalikan 500 dengan galat Postgres
"inconsistent types deduced for parameter $2".

Penyebab: pada UPDATE hris.loans, $2 dipakai sekaligus sebagai nilai kolom
numeric (remaining_balance = $2) DAN dibandingkan dengan literal integer
($2 <= 0 di dua ekspresi CASE). Postgres tidak bisa menyimpulkan satu tipe
untuk parameter yang sama, lalu menolak seluruh pernyataan.

Bug ini dorman sejak ditulis: blok alokasi pinjaman hanya dieksekusi bila
ada payroll_details dengan loan_deduction > 0. Sampai Juli tidak ada
karyawan yang cicilannya jatuh tempo, sehingga baru muncul pada periode
Agustus ketika cicilan pertama mulai berjalan.

Perbaikan: cast eksplisit $2::numeric dan $3::numeric.

Tidak ada data yang rusak — transaksi sudah rollback dengan benar, run
Agustus tetap berstatus completed dan saldo pinjaman utuh (5.000.001,
paid_amount 0). Diverifikasi dengan menjalankan SQL yang diperbaiki
terhadap baris pinjaman sungguhan di dalam transaksi lalu di-rollback:
angsuran biasa menghasilkan sisa 3.333.334, dan pelunasan menghasilkan
sisa 0 dengan status paid_off serta is_active false. Penyisiran pola serupa
(parameter dipakai sebagai nilai kolom sekaligus dibanding literal) tidak
menemukan kasus lain.

+++ src/app/api/hris/payroll/[id]/route.ts
@@ -333,11 +333,15 @@ async function markRunPaidAndSettleLoans(
 
       for (const alloc of allocations) {
         await client.query(
+          // $2 di-cast eksplisit: dipakai sekaligus sebagai nilai kolom
+          // numeric DAN dibandingkan dengan literal 0, sehingga tanpa cast
+          // Postgres menolak dengan "inconsistent types deduced for
+          // parameter $2".
           `UPDATE hris.loans
-           SET remaining_balance = $2,
-               paid_amount = COALESCE(paid_amount, 0) + $3,
-               is_active = CASE WHEN $2 <= 0 THEN false ELSE is_active END,
-               status = CASE WHEN $2 <= 0 THEN 'paid_off' ELSE status END,
+           SET remaining_balance = $2::numeric,
+               paid_amount = COALESCE(paid_amount, 0) + $3::numeric,
+               is_active = CASE WHEN $2::numeric <= 0 THEN false ELSE is_active END,
+               status = CASE WHEN $2::numeric <= 0 THEN 'paid_off' ELSE status END,
                updated_at = now()
            WHERE id = $1`,
           [alloc.loanId, alloc.newRemaining, alloc.amount]
